Address tim's and brandon's comments
